Add useResumeCount to State Transitions stories
